+++ CHANGELOG.rst
@@ -1,3 +1,23 @@
+0.6.0
+-----
+
+- Now ``QtAsyncRunner.close()`` will cause all running coroutines to never be called back into the main thread.
+
+  Letting coroutines resume into the main thread after ``close()``
+  has been called can be problematic, specially in tests, as ``close()`` is often called during test teardown.
+  If the user missed to properly wait on a coroutine during the test, what can
+  happen is that the coroutine will resume (when the thread finishes), possibly after resources have already
+  been cleared, specially widgets.
+
+  Dropping seems harsh, but follows what other libraries like ``asyncio`` do when faced with the same
+  situation.
+
+  We might consider adding a ``wait`` flag or something like that in the future to instead of cancelling the coroutines,
+  wait for them.
+
+- ``AsyncTester.start_and_wait()`` now receives an optional ``timeout_s`` parameter, which overwrites
+  ``AsyncTester.timeout_s``.
+
 0.5.2
 -----